+++ Fabric/src/main/java/net/merchantpug/apugli/mixin/fabric/common/EnchantmentHelperMixin.java
@@ -1,11 +1,8 @@
 package net.merchantpug.apugli.mixin.fabric.common;
 
 import com.llamalad7.mixinextras.injector.ModifyExpressionValue;
-import net.merchantpug.apugli.access.ItemStackAccess;
-import net.merchantpug.apugli.platform.Services;
 import net.merchantpug.apugli.registry.power.ApugliPowers;
 import net.minecraft.nbt.ListTag;
-import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.enchantment.Enchantment;
 import net.minecraft.world.item.enchantment.EnchantmentHelper;
@@ -13,7 +10,6 @@ import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.ModifyVariable;
-import org.spongepowered.asm.mixin.injection.Redirect;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 
@@ -52,13 +48,6 @@ public abstract class EnchantmentHelperMixin {
     @ModifyVariable(method = "getItemEnchantmentLevel", at = @At(value = "INVOKE_ASSIGN", target = "Lnet/minecraft/world/item/ItemStack;getEnchantmentTags()Lnet/minecraft/nbt/ListTag;"))
     private static ListTag getEnchantmentsGetLevel(ListTag original) {
         return ApugliPowers.MODIFY_ENCHANTMENT_LEVEL.get().getEnchantments(apugli$itemEnchantmentLevelStack, original);
-    }
-
-    @Inject(method = "getEnchantmentLevel(Lnet/minecraft/world/item/enchantment/Enchantment;Lnet/minecraft/world/entity/LivingEntity;)I", at = @At("RETURN"), cancellable = true)
-    private static void getEquipmentLevel(Enchantment enchantment, LivingEntity entity, CallbackInfoReturnable<Integer> cir) {
-        int originalReturn = cir.getReturnValue();
-        int newEnchantLevel = (int) Services.PLATFORM.applyModifiers(entity, ApugliPowers.MODIFY_ENCHANTMENT_LEVEL.get(), originalReturn, power -> ApugliPowers.MODIFY_ENCHANTMENT_LEVEL.get().doesApply(power, enchantment));
-        cir.setReturnValue(newEnchantLevel);
     }
 
 }